+++ haze-blur/src/commonTest/kotlin/dev/chrisbanes/haze/blur/BlurVisualEffectLifecycleTest.kt
@@ -16,6 +16,7 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isFalse
 import assertk.assertions.isTrue
+import dev.chrisbanes.haze.ExperimentalHazeApi
 import dev.chrisbanes.haze.HazeArea
 import dev.chrisbanes.haze.HazeInputScale
 import dev.chrisbanes.haze.HazeState
@@ -26,6 +27,7 @@ import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.test.Test
 import kotlinx.coroutines.CoroutineScope
 
+@OptIn(ExperimentalHazeApi::class)
 class BlurVisualEffectLifecycleTest {
 
   @Test