+++ haze/src/commonTest/kotlin/dev/chrisbanes/haze/HazeCoordinatesTest.kt
@@ -11,6 +11,7 @@ import assertk.assertions.isEqualTo
 import assertk.assertions.isNull
 import kotlin.test.Test
 
+@OptIn(ExperimentalHazeApi::class)
 class HazeCoordinatesTest {
 
   @Test